+++ deflect/StreamPrivate.h
@@ -51,8 +51,7 @@
 #include "Stream.h" // Stream::Future
 
 #include <string>
-
-class QString;
+#include <memory>
 
 namespace deflect
 {
@@ -62,10 +61,8 @@ class StreamSendWorker;
 /**
  * Private implementation for the Stream class.
  */
-class StreamPrivate : public QObject
+class StreamPrivate
 {
-    Q_OBJECT
-
 public:
     /**
      * Create a new stream and open a new connection to the deflect::Server.
@@ -74,16 +71,21 @@ public:
      * e.g. "192.168.1.83" This method must be called by all Streams sharing a
      * common identifier before any of them starts sending images.
      *
-     * @param stream the parent object owning this object
      * @param name the unique stream name
      * @param address Address of the target Server instance.
      * @param port Port of the target Server instance.
      */
-    StreamPrivate( Stream* stream, const std::string& name,
-                   const std::string& address, const unsigned short port );
+    StreamPrivate( const std::string& name, const std::string& address,
+                   const unsigned short port );
 
     /** Destructor, close the Stream. */
     ~StreamPrivate();
+
+    /** Send the open message to the server. */
+    void sendOpen();
+
+    /** Send the quit message to the server. */
+    void sendClose();
 
     /**
      * Close the stream.
@@ -124,12 +126,8 @@ public:
     /** Has a successful event registration reply been received */
     bool registeredForEvents;
 
-private slots:
-    void _onDisconnected();
-
 private:
-    Stream* _parent;
-    StreamSendWorker* _sendWorker;
+    std::unique_ptr< StreamSendWorker > _sendWorker;
 };
 
 }